requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
langchain-groq==0.1.0
langchain-community==0.1.2
langchain-core==0.0.3
langchain-chains==0.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
